add functionality to sort by semester
